grizzly/requirements.txt
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.6
colorama==0.4.4
Django==3.1.7
djangorestframework==3.12.2
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mysqlclient==2.0.3
pep8==1.7.1
Pillow==8.1.2
pycodestyle==2.7.0
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
